requirement.txt
# core
pyspark == 3.5.3

# Data handling
pandas==2.2.2
pyarrow==16.1.0   # Needed for Parquet/Arrow conversion

# Utils
python-dotenv==1.0.1   # For managing env variables later
ruamel.yaml==0.18.6    # For configs/schemas (better YAML support)

# Optional but useful
jupyterlab==4.2.3      # For notebooks/exploration
notebook==7.2.1